tda/requirements.txt
pytest==7.2.1
pytest-xdist==3.1.0
pytest-forked==1.4.0
urllib3==1.26.14
selenium==4.8.0
Appium-Python-Client==2.8.1
PyYAML==6.0
requests==2.28.2
sentry-sdk==1.14.0
